// src/app/layout/trcm-lab-automation/models/repair/delivery.ts
export class Delivery {

    id: number;
  intrash: string;
  location: string;
  serialNumber: string;
    devices: {
      action: '';
      actionStatus: '';
      contract_expires: Date;
      contract_starts: Date;
      contractperiod: number;
      creationDate: '';
      deviceContractStatus: '';
      deviceModels: '';
      deviceWarantyStatus: '';
      deviceowner: '';
      id: number;
      imeinumber: '';
      intrash: '';
      manufacturer: '';
      partnumber: '';
      seller: '';
      serialnumber: '';
      warrantyExpire: Date;
      warranty_starts: Date;
      warrantyperiod: number;
    };
    action: string;
  actionStatus: string;
  creationOn: Date;
  customers: string;
  deliveredBy: string;
  deliveryStatus: string;

  }